+++ corvid/table_aggregation/pairwise_mapping.py
